+++ NEWS.md
@@ -1,3 +1,12 @@
+# 0.0.2 (2017-10-11)
+
+### Breaking changes
+### New functions / parameters
+- analyze.merModLmerTest
+### Major changes
+### Minor changes
+
+
 # [0.0.1](https://github.com/neuropsychology/psycho.R/releases/tag/0.0.1) (2017-10-06)
 
 ### Breaking changes